+++ src/factorioModRuntime.ts
@@ -312,6 +312,11 @@ export class FactorioModRuntime extends EventEmitter {
 						this.updateBreakpoints();
 					}
 					this.sendEvent('stopOnBreakpoint');
+				} else if (event.startsWith("exception")) {
+					// notify stop on exception
+					const err = event.substr(10)
+					this.sendEvent('output', err, "stderr");
+					this.sendEvent('stopOnException', err);
 				} else {
 					// unexpected event?
 					this.output.appendLine("unexpected event: " + event);